+ dummy code cleaning - remove all PlotDataProviders except MonitoringDataProvider
